chore: alpha.2 release

+++ server/src/main/java/datart/server/base/params/FolderCreateParam.java
@@ -30,6 +30,6 @@ public class FolderCreateParam extends VizCreateParam {
 
     private String parentId;
 
-    private int index;
+    private double index;
 
 }
